// src/app/core/interfaces/documento.ts
export interface Documento{
    idDocumento:    number;
    idGrupo:        number;
    nombreGrupo:    string;
    idCia:          number;
    titulo:         string;
    descripcion:    string;
    codMunicipio:   string;
    nombreArchivo:  string;
    archivo:        string;
    activo:         boolean;    
}